requirements.txt
#Admin LTE for frontend templates
#
psycopg2
openpyxl
pandas
django-crispy-forms
django-bootstrap-icons


